requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
pytesseract
pdfplumber
python-docx
transformers